Add sphinx.util.logging.SafeEncodingWriter
